requirements_2.txt
flask==3.0.0
requests==2.31.0
google-generativeai==0.8.3
httpx==0.27.0
gunicorn==21.2.0
